# Optimise one of the org dashboard queries