Ticket: https://trello.com/c/V6vkvxon/254-review-loading-time-for-organisation-page

This is the first in a series of small, incremental improvements to the org dashboard page. This PR focuses on optimising one of the slower queries. This specific query adds a flat latency to the page regardless of which organisation you're viewing, so for small orgs this query takes up a disproportionate amount of the page load. Optimising this query will provide a significant relative increase in page loading for smaller orgs.

**Note**: query plans/timings are from our semi-representative large staging DB

## Baseline

<details><summary>Baseline - SQL</summary>

```sql
 SELECT organisation.name               AS organisation_name,
        organisation.id                 AS organisation_id,
        services.name                   AS service_name,
        services.id                     AS service_id,
        max(anon_1.free_allowance)      AS free_allowance,
        min(anon_1.free_allowance_left) AS free_allowance_left,
        sum(anon_1.chargeable_units)    AS chargeable_units,
        sum(anon_1.charged_units)       AS charged_units,
        sum(anon_1.cost)                AS cost,
        services.active
 FROM services
          LEFT OUTER JOIN (SELECT services.id                                                         AS service_id,
                                  ft_billing.bst_date                                                 AS bst_date,
                                  annual_billing.free_sms_fragment_limit                              AS free_allowance,
                                  greatest(greatest(annual_billing.free_sms_fragment_limit - coalesce(
                                          CAST(sum(coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0))
                                               OVER (PARTITION BY ft_billing.service_id ORDER BY ft_billing.bst_date ROWS BETWEEN UNBOUNDED PRECEDING AND 1 PRECEDING) AS INTEGER),
                                          0), 0) - coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0),
                                           0)                                                         AS free_allowance_left,
                                  coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0) AS chargeable_units,
                                  greatest(coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0) -
                                           greatest(annual_billing.free_sms_fragment_limit - coalesce(CAST(sum(
                                                                                                           coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0))
                                                                                                           OVER (PARTITION BY ft_billing.service_id ORDER BY ft_billing.bst_date ROWS BETWEEN UNBOUNDED PRECEDING AND 1 PRECEDING) AS INTEGER),
                                                                                                      0), 0), 0) *
                                  coalesce(ft_billing.rate, 0.0)                                      AS cost,
                                  greatest(coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0) -
                                           greatest(annual_billing.free_sms_fragment_limit - coalesce(CAST(sum(
                                                                                                           coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0))
                                                                                                           OVER (PARTITION BY ft_billing.service_id ORDER BY ft_billing.bst_date ROWS BETWEEN UNBOUNDED PRECEDING AND 1 PRECEDING) AS INTEGER),
                                                                                                      0), 0),
                                           0)                                                         AS charged_units
                           FROM services
                                    LEFT OUTER JOIN annual_billing ON annual_billing.service_id = services.id AND
                                                                      annual_billing.financial_year_start = 2022
                                    LEFT OUTER JOIN ft_billing ON services.id = ft_billing.service_id AND
                                                                  ft_billing.bst_date >= '2022-04-01' AND
                                                                  ft_billing.bst_date <= '2023-03-31' AND
                                                                  ft_billing.notification_type = 'sms') AS anon_1
                          ON services.id = anon_1.service_id
          LEFT OUTER JOIN organisation ON organisation.id = services.organisation_id
 WHERE anon_1.bst_date >= '2022-04-01'
   AND anon_1.bst_date <= '2023-03-31'
   AND services.organisation_id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::UUID
   AND services.restricted IS false
 GROUP BY organisation.name, organisation.id, services.id, services.name
 ORDER BY organisation.name, services.name;
```
</details>
<details><summary>Baseline - Plan</summary>

```
 Sort  (cost=282586.21..282586.53 rows=129 width=141) (actual time=1949.972..1949.982 rows=31 loops=1)
   Sort Key: organisation.name, services.name
   Sort Method: quicksort  Memory: 29kB
   ->  GroupAggregate  (cost=282577.50..282581.69 rows=129 width=141) (actual time=1947.203..1949.935 rows=31 loops=1)
         Group Key: organisation.id, services.id
         ->  Sort  (cost=282577.50..282577.82 rows=129 width=133) (actual time=1947.127..1947.676 rows=11774 loops=1)
               Sort Key: organisation.id, services.id
               Sort Method: quicksort  Memory: 2040kB
               ->  Hash Left Join  (cost=206586.92..282572.97 rows=129 width=133) (actual time=1129.299..1943.096 rows=11774 loops=1)
                     Hash Cond: (services.organisation_id = organisation.id)
                     ->  Hash Join  (cost=206578.61..282562.89 rows=129 width=105) (actual time=1129.255..1940.738 rows=11774 loops=1)
                           Hash Cond: (anon_1.service_id = services.id)
                           ->  Subquery Scan on anon_1  (cost=204002.84..279350.84 rows=242391 width=64) (actual time=1112.211..1876.579 rows=834803 loops=1)
                                 Filter: ((anon_1.bst_date >= '2022-04-01'::date) AND (anon_1.bst_date <= '2023-03-31'::date))
                                 Rows Removed by Filter: 17790
                                 ->  WindowAgg  (cost=204002.84..266792.84 rows=837200 width=84) (actual time=1112.210..1799.328 rows=852593 loops=1)
                                       ->  Sort  (cost=204002.84..206095.84 rows=837200 width=51) (actual time=1112.188..1230.604 rows=852593 loops=1)
                                             Sort Key: ft_billing.service_id, ft_billing.bst_date
                                             Sort Method: external merge  Disk: 52040kB
                                             ->  Hash Left Join  (cost=4933.11..64410.40 rows=837200 width=51) (actual time=17.093..665.850 rows=852593 loops=1)
                                                   Hash Cond: (services_1.id = annual_billing.service_id)
                                                   ->  Hash Right Join  (cost=3198.33..60477.56 rows=837200 width=47) (actual time=8.016..492.546 rows=852593 loops=1)
                                                         Hash Cond: (ft_billing.service_id = services_1.id)
                                                         ->  Index Scan using ix_ft_billing_bst_date_sms_only on ft_billing  (cost=0.43..55081.60 rows=837200 width=31) (actual time=0.019..317.871 rows=834803 loops=1)
                                                               Index Cond: ((bst_date >= '2022-04-01'::date) AND (bst_date <= '2023-03-31'::date))
                                                         ->  Hash  (cost=2915.18..2915.18 rows=22618 width=16) (actual time=7.956..7.957 rows=22618 loops=1)
                                                               Buckets: 32768  Batches: 1  Memory Usage: 1317kB
                                                               ->  Seq Scan on services services_1  (cost=0.00..2915.18 rows=22618 width=16) (actual time=0.005..5.421 rows=22618 loops=1)
                                                   ->  Hash  (cost=1473.49..1473.49 rows=20903 width=20) (actual time=9.000..9.001 rows=21034 loops=1)
                                                         Buckets: 32768  Batches: 1  Memory Usage: 1325kB
                                                         ->  Seq Scan on annual_billing  (cost=0.00..1473.49 rows=20903 width=20) (actual time=0.954..5.805 rows=21034 loops=1)
                                                               Filter: (financial_year_start = 2022)
                                                               Rows Removed by Filter: 46285
                           ->  Hash  (cost=2570.22..2570.22 rows=444 width=57) (actual time=1.098..1.099 rows=260 loops=1)
                                 Buckets: 1024  Batches: 1  Memory Usage: 32kB
                                 ->  Bitmap Heap Scan on services  (cost=223.32..2570.22 rows=444 width=57) (actual time=0.182..1.050 rows=260 loops=1)
                                       Recheck Cond: (organisation_id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::uuid)
                                       Filter: (restricted IS FALSE)
                                       Rows Removed by Filter: 1180
                                       Heap Blocks: exact=633
                                       ->  Bitmap Index Scan on ix_services_organisation_id  (cost=0.00..223.21 rows=1440 width=0) (actual tim
e=0.107..0.107 rows=1440 loops=1)
                                             Index Cond: (organisation_id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::uuid)
                     ->  Hash  (cost=8.30..8.30 rows=1 width=44) (actual time=0.024..0.025 rows=1 loops=1)
                           Buckets: 1024  Batches: 1  Memory Usage: 9kB
                           ->  Index Scan using organisation_pkey on organisation  (cost=0.28..8.30 rows=1 width=44) (actual time=0.020..0.021 rows=1 loops=1)
                                 Index Cond: (id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::uuid)
 Planning Time: 0.615 ms
 Execution Time: 1957.183 ms
(48 rows)
```
</details>

##  Commit 1: Passing organisation_id into subquery

This function collates billing data for services in a given organisation. It calls 
`fetch_usage_for_all_services_sms` and applies a filter for live services in the specified organisation. This would 
be ok if not for the fact that `fetch_usage_for_all_services_sms` builds a subquery using 
`_fetch_usage_for_all_services_sms_query`. The subquery doesn't get 
filtered, and so it ends up first running a query joining all services to all ft_billing data. This is way more data 
than we need to be querying when looking up billing for a single organisation. We currently emit a query like this:

This patch feeds an optional `organisation_id` through the relevant functions so that we can finally end up with a 
single extra filter clause and a query like this:

<details><summary>Commit 1 - SQL</summary>

```sql
SELECT organisation.name               AS organisation_name,
        organisation.id                 AS organisation_id,
        services.name                   AS service_name,
        services.id                     AS service_id,
        max(anon_1.free_allowance)      AS free_allowance,
        min(anon_1.free_allowance_left) AS free_allowance_left,
        sum(anon_1.chargeable_units)    AS chargeable_units,
        sum(anon_1.charged_units)       AS charged_units,
        sum(anon_1.cost)                AS cost,
        services.active
 FROM services
          LEFT OUTER JOIN (SELECT services.id                                                         AS service_id,
                                  ft_billing.bst_date                                                 AS bst_date,
                                  annual_billing.free_sms_fragment_limit                              AS free_allowance,
                                  greatest(greatest(annual_billing.free_sms_fragment_limit - coalesce(
                                          CAST(sum(coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0))
                                               OVER (PARTITION BY ft_billing.service_id ORDER BY ft_billing.bst_date ROWS BETWEEN UNBOUNDED PRECEDING AND 1 PRECEDING) AS INTEGER),
                                          0), 0) - coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0),
                                           0)                                                         AS free_allowance_left,
                                  coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0) AS chargeable_units,
                                  greatest(coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0) -
                                           greatest(annual_billing.free_sms_fragment_limit - coalesce(CAST(sum(
                                                                                                           coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0))
                                                                                                           OVER (PARTITION BY ft_billing.service_id ORDER BY ft_billing.bst_date ROWS BETWEEN UNBOUNDED PRECEDING AND 1 PRECEDING) AS INTEGER),
                                                                                                      0), 0), 0) *
                                  coalesce(ft_billing.rate, 0.0)                                      AS cost,
                                  greatest(coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0) -
                                           greatest(annual_billing.free_sms_fragment_limit - coalesce(CAST(sum(
                                                                                                           coalesce(ft_billing.billable_units * ft_billing.rate_multiplier, 0))
                                                                                                           OVER (PARTITION BY ft_billing.service_id ORDER BY ft_billing.bst_date ROWS BETWEEN UNBOUNDED PRECEDING AND 1 PRECEDING) AS INTEGER),
                                                                                                      0), 0),
                                           0)                                                         AS charged_units
                           FROM services
                                    LEFT OUTER JOIN annual_billing ON annual_billing.service_id = services.id AND
                                                                      annual_billing.financial_year_start = 2022
                                    LEFT OUTER JOIN ft_billing ON services.id = ft_billing.service_id AND
                                                                  ft_billing.bst_date >= '2022-04-01' AND
                                                                  ft_billing.bst_date <= '2023-03-31' AND
                                                                  ft_billing.notification_type = 'sms'

-- the new bit: start
                           WHERE services.organisation_id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::UUID) AS anon_1
-- the new bit: end

                          ON services.id = anon_1.service_id
          LEFT OUTER JOIN organisation ON organisation.id = services.organisation_id
 WHERE anon_1.bst_date >= '2022-04-01'
   AND anon_1.bst_date <= '2023-03-31'
   AND services.organisation_id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::UUID
   AND services.restricted IS false
 GROUP BY organisation.name, organisation.id, services.id, services.name
 ORDER BY organisation.name, services.name;
```
</details>

<details><summary>Commit 1 - Plan</summary>

```
 Sort  (cost=59376.90..59376.91 rows=1 width=141) (actual time=330.758..330.763 rows=5 loops=1)
   Sort Key: organisation.name, services.name
   Sort Method: quicksort  Memory: 26kB
   ->  GroupAggregate  (cost=59376.86..59376.89 rows=1 width=141) (actual time=330.612..330.753 rows=5 loops=1)
         Group Key: organisation.id, services.id
         ->  Sort  (cost=59376.86..59376.86 rows=1 width=133) (actual time=330.560..330.598 rows=723 loops=1)
               Sort Key: organisation.id, services.id
               Sort Method: quicksort  Memory: 198kB
               ->  Nested Loop Left Join  (cost=59188.56..59376.85 rows=1 width=133) (actual time=328.652..330.358 rows=723 loops=1)
                     ->  Hash Join  (cost=59188.28..59369.71 rows=1 width=105) (actual time=328.639..329.511 rows=723 loops=1)
                           Hash Cond: (anon_1.service_id = services.id)
                           ->  Subquery Scan on anon_1  (cost=58985.53..59165.44 rows=579 width=64) (actual time=328.535..329.258 rows=743 loops=1)
                                 Filter: ((anon_1.bst_date >= '2022-04-01'::date) AND (anon_1.bst_date <= '2023-03-31'::date))
                                 Rows Removed by Filter: 47
                                 ->  WindowAgg  (cost=58985.53..59135.46 rows=1999 width=84) (actual time=328.534..329.174 rows=790 loops=1)
                                       ->  Sort  (cost=58985.53..58990.53 rows=1999 width=51) (actual time=328.520..328.567 rows=790 loops=1)
                                             Sort Key: ft_billing.service_id, ft_billing.bst_date
                                             Sort Method: quicksort  Memory: 133kB
                                             ->  Hash Right Join  (cost=635.28..58875.94 rows=1999 width=51) (actual time=0.342..328.085 rows=790 loops=1)
                                                   Hash Cond: (ft_billing.service_id = services_1.id)
                                                   ->  Index Scan using ix_ft_billing_bst_date_sms_only on ft_billing  (cost=0.43..55081.60 rows=837200 width=31) (actual time=0.014..270.877 rows=834803 loops=1)
                                                         Index Cond: ((bst_date >= '2022-04-01'::date) AND (bst_date <= '2023-03-31'::date))
                                                   ->  Hash  (cost=634.17..634.17 rows=54 width=20) (actual time=0.225..0.226 rows=54 loops=1)
                                                         Buckets: 1024  Batches: 1  Memory Usage: 11kB
                                                         ->  Nested Loop Left Join  (cost=9.25..634.17 rows=54 width=20) (actual time=0.018..0.216 rows=54 loops=1)
                                                               ->  Bitmap Heap Scan on services services_1  (cost=8.83..202.55 rows=54 width=16) (actual time=0.010..0.034 rows=54 loops=1)
                                                                     Recheck Cond: (organisation_id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::uuid)
                                                                     Heap Blocks: exact=47
                                                                     ->  Bitmap Index Scan on ix_services_organisation_id  (cost=0.00..8.82 rows=54 width=0) (actual time=0.006..0.006 rows=54 loops=1)
                                                                           Index Cond: (organisation_id = 'ee0e5030-3a7b-4da8-8685-b6efb770231f'::uuid)
…