@samuelhwilliams 2023-03-31
Good reminder; I've added the timings to the 2nd commit.

I will merge this on Monday 🙃